split up files, add new tests
