+++ .github/workflows/run-integration-tests.yml
@@ -36,4 +36,4 @@ jobs:
 
     - name: Run integration tests
       run: |
-        ./.venv/bin/pytest -n auto -m "integration and not performance"
+        ./.venv/bin/pytest -n auto -m "integration and not performance" -vv -s --tb=short --log-cli-level=INFO --capture=no
