+++ tests/integration/test_deterministic_behavior.py
@@ -1,0 +1,114 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+"""Tests for deterministic behavior with random seeds."""
+
+import pytest
+
+from tests.integration.conftest import AIPerfCLI
+from tests.integration.models import FakeAIServer
+
+
+@pytest.mark.integration
+@pytest.mark.asyncio
+class TestDeterministicBehavior:
+    """Tests for deterministic behavior with random seeds."""
+
+    async def test_same_seed_identical_inputs(
+        self, cli: AIPerfCLI, fakeai_server: FakeAIServer
+    ):
+        """Same random seed produces identical payloads across runs."""
+        result1 = await cli.run(
+            f"""
+            aiperf profile \
+                --model openai/gpt-oss-20b \
+                --url {fakeai_server.url} \
+                --endpoint-type chat \
+                --request-count 10 \
+                --concurrency 2 \
+                --random-seed 42 \
+                --image-width-mean 64 \
+                --image-height-mean 64 \
+                --audio-length-mean 0.1 \
+                --workers-max 1 \
+                --ui simple
+            """
+        )
+
+        result2 = await cli.run(
+            f"""
+            aiperf profile \
+                --model openai/gpt-oss-20b \
+                --url {fakeai_server.url} \
+                --endpoint-type chat \
+                --request-count 10 \
+                --concurrency 2 \
+                --random-seed 42 \
+                --image-width-mean 64 \
+                --image-height-mean 64 \
+                --audio-length-mean 0.1 \
+                --workers-max 1 \
+                --ui simple
+            """
+        )
+
+        assert result1.request_count == 10
+        assert result2.request_count == 10
+
+        inputs_1 = result1.inputs.data
+        inputs_2 = result2.inputs.data
+
+        assert len(inputs_1) == len(inputs_2), "Session counts differ"
+
+        for s1, s2 in zip(inputs_1, inputs_2, strict=True):
+            assert s1.session_id != s2.session_id
+            assert s1.payloads == s2.payloads
+
+    async def test_different_seeds_different_inputs(
+        self, cli: AIPerfCLI, fakeai_server: FakeAIServer
+    ):
+        """Different random seeds produce different payloads."""
+        result1 = await cli.run(
+            f"""
+            aiperf profile \
+                --model openai/gpt-oss-20b \
+                --url {fakeai_server.url} \
+                --endpoint-type chat \
+                --request-count 10 \
+                --concurrency 2 \
+                --random-seed 42 \
+                --image-width-mean 128 \
+                --image-height-mean 128 \
+                --workers-max 1 \
+                --ui simple
+            """
+        )
+
+        result2 = await cli.run(
+            f"""
+            aiperf profile \
+                --model openai/gpt-oss-20b \
+                --url {fakeai_server.url} \
+                --endpoint-type chat \
+                --request-count 10 \
+                --concurrency 2 \
+                --random-seed 123 \
+                --image-width-mean 128 \
+                --image-height-mean 128 \
+                --workers-max 1 \
+                --ui simple
+            """
+        )
+
+        assert result1.request_count == 10
+        assert result2.request_count == 10
+
+        inputs_1 = result1.inputs.data
+        inputs_2 = result2.inputs.data
+
+        payloads_different = False
+        for s1, s2 in zip(inputs_1, inputs_2, strict=True):
+            if s1.payloads != s2.payloads:
+                payloads_different = True
+                break
+
+        assert payloads_different, "Different seeds should produce different payloads"
